Wed, Oct 18, 2023, 10:12 AM +02:00

+++ filter-less/helpers.c
@@ -9,7 +9,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            int RGBaverage = round(image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
+            int RGBaverage = (BYTE)round(image[i][j].rgbtRed + image[i][j].rgbtGreen + image[i][j].rgbtBlue) / 3.0;
 
             image[i][j].rgbtRed = RGBaverage;
             image[i][j].rgbtGreen = RGBaverage;
